Enforce some basic post policies

+++ app/posts/services.py
@@ -7,13 +7,14 @@ from app.users.models import User
 
 from .exceptions import ForbiddenException
 from .models import CreatePostData, Post, UpdatePostData
+from .policies import PostPolicy
 from .repositories import PostRepositoryABC
 
 
 class PostServiceABC(ABC):
     @abstractmethod
     def get_posts(
-        self, page: int, size: int, search: str | None, db: Session
+        self, page: int, size: int, search: str | None, db: Session, current_user: User
     ) -> Sequence[tuple[Post, int]]:
         pass
 
@@ -24,7 +25,9 @@ class PostServiceABC(ABC):
         pass
 
     @abstractmethod
-    def get_post_by_id(self, post_id: str, db: Session) -> Post | None:
+    def get_post_by_id(
+        self, post_id: str, db: Session, current_user: User
+    ) -> Post | None:
         pass
 
     @abstractmethod
@@ -39,17 +42,27 @@ class PostServiceABC(ABC):
 
 
 class PostService(PostServiceABC):
-    def __init__(self, post_repository: PostRepositoryABC):
+    def __init__(self, post_repository: PostRepositoryABC, post_policy: PostPolicy):
         self._post_repository = post_repository
+        self._post_policy = post_policy
 
     def get_posts(
-        self, page: int, size: int, search: str | None, db: Session
+        self, page: int, size: int, search: str | None, db: Session, current_user: User
     ) -> Sequence[tuple[Post, int]]:
-        return self._post_repository.get_posts(page, size, search, db)
+        posts = self._post_repository.get_posts(page, size, search, db)
+        filtered_posts = [
+            (post, votes)
+            for post, votes in posts
+            if self._post_policy.can_view(post, current_user)
+        ]
+        return filtered_posts
 
     def create_post(
         self, post_data: CreatePostData, db: Session, current_user: User
     ) -> Post:
+        if not self._post_policy.can_create(current_user):
+            raise ForbiddenException()
+
         new_post = Post(
             owner_id=current_user.id,
             title=post_data.title,
@@ -59,8 +72,15 @@ class PostService(PostServiceABC):
         )
         return self._post_repository.create_post(new_post, db)
 
-    def get_post_by_id(self, post_id: str, db: Session) -> Post | None:
-        return self._post_repository.get_post_by_id(post_id, db)
+    def get_post_by_id(
+        self, post_id: str, db: Session, current_user: User
+    ) -> Post | None:
+        post = self._post_repository.get_post_by_id(post_id, db)
+
+        if not self._post_policy.can_view(post, current_user):
+            raise ForbiddenException()
+
+        return post
 
     def update_post(
         self, post_id: str, post_data: UpdatePostData, db: Session, current_user: User
@@ -68,7 +88,8 @@ class PostService(PostServiceABC):
         post = self._post_repository.get_post_by_id(post_id, db)
         if not post:
             return None
-        self._ensure_owner(post, current_user)
+        if not self._post_policy.can_update(post, current_user):
+            raise ForbiddenException()
         update_data = self._build_update_data(post_data)
         return self._post_repository.update_post(post, update_data, db)
 
@@ -76,7 +97,8 @@ class PostService(PostServiceABC):
         post = self._post_repository.get_post_by_id(post_id, db)
         if not post:
             return None
-        self._ensure_owner(post, current_user)
+        if not self._post_policy.can_delete(post, current_user):
+            raise ForbiddenException()
         self._post_repository.delete_post(post, db)
         return None
 
@@ -91,7 +113,3 @@ class PostService(PostServiceABC):
         if post_data.rating is not None:
             update_data[getattr(Post, "rating")] = post_data.rating
         return update_data
-
-    def _ensure_owner(self, post: Post, current_user: User):
-        if getattr(post, "owner_id", None) != current_user.id:
-            raise ForbiddenException()
